pyproject.toml
[project]
name = "yt-answer-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyterlab>=4.5.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "youtube-transcript-api>=1.2.3",
]
